+++ packages/gestalt/src/Masonry.js
@@ -55,8 +55,6 @@ type Props<T> = {|
   virtualBoundsTop?: number,
   virtualBoundsBottom?: number,
   virtualize?: boolean,
-
-  fixFetchMoreBug?: boolean,
 |};
 
 type State<T> = {|
@@ -190,11 +188,6 @@ export default class Masonry<T: {}> extends React.Component<
      * Whether or not to use actual virtualization
      */
     virtualize: PropTypes.bool,
-
-    /**
-     * Flag to decide if we should fix fetch more bug (see commit notes)
-     */
-    fixFetchMoreBug: PropTypes.bool,
   };
 
   static defaultProps = {
@@ -203,8 +196,6 @@ export default class Masonry<T: {}> extends React.Component<
     layout: DefaultLayoutSymbol,
     loadItems: () => {},
     virtualize: false,
-
-    fixFetchMoreBug: false,
   };
 
   constructor(props: Props<T>) {
@@ -462,7 +453,6 @@ export default class Masonry<T: {}> extends React.Component<
       gutterWidth: gutter,
       items,
       minCols,
-      fixFetchMoreBug,
     } = this.props;
     const { hasPendingMeasurements, measurementStore, width } = this.state;
 
@@ -601,9 +591,7 @@ export default class Masonry<T: {}> extends React.Component<
               isFetching={
                 this.state.isFetching || this.state.hasPendingMeasurements
               }
-              scrollHeight={
-                height + (fixFetchMoreBug ? this.containerOffset : 0)
-              }
+              scrollHeight={height + this.containerOffset}
               scrollTop={this.state.scrollTop}
             />
           )}